Rename audio recording details file
