+++ src/app/component/admin/audio-recordings/audio-recordings.module.ts
@@ -3,7 +3,7 @@ import { RouterModule } from "@angular/router";
 import { GetRouteConfigForPage } from "@helpers/page/pageRouting";
 import { SharedModule } from "@shared/shared.module";
 import { adminAudioRecordingsRoute } from "./audio-recordings.menus";
-import { AdminAudioRecordingComponent } from "./detail/detail.component";
+import { AdminAudioRecordingComponent } from "./details/details.component";
 import { AdminAudioRecordingsComponent } from "./list/list.component";
 
 const components = [
