+++ src/app/component/shared/detail-view/render-field/render-field.component.ts
@@ -1,6 +1,12 @@
-import { ChangeDetectorRef, Component, Input, OnInit } from "@angular/core";
+import {
+  ChangeDetectorRef,
+  Component,
+  Input,
+  OnInit,
+  OnChanges,
+} from "@angular/core";
 import { WithUnsubscribe } from "@helpers/unsubscribe/unsubscribe";
-import { AbstractModel } from "@models/AbstractModel";
+import { AbstractModel, isResolvedModel } from "@models/AbstractModel";
 import { DateTime, Duration } from "luxon";
 import { Observable } from "rxjs";
 import { takeUntil } from "rxjs/operators";
@@ -81,9 +87,7 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
     } else if (value instanceof Observable) {
       this.humanizeObservable(value);
     } else if (value instanceof AbstractModel) {
-      this.styling = FieldStyling.Model;
-      this.display = "";
-      this.model = value;
+      this.humanizeAbstractModel(value);
     } else if (typeof value === "object") {
       // TODO Implement optional treeview
       this.humanizeObject(value);
@@ -98,9 +102,23 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
   }
 
   /**
+   * Convert abstract model to human readable output
+   * @param value Display input
+   */
+  private humanizeAbstractModel(value: AbstractModel) {
+    if (isResolvedModel(value)) {
+      this.styling = FieldStyling.Model;
+      this.display = "";
+      this.model = value;
+    } else {
+      this.setLoading();
+    }
+  }
+
+  /**
    * Convert string to human readable output. Currently this only checks if the
    * string is an image url.
-   * @param value Display output
+   * @param value Display input
    */
   private humanizeString(value: string) {
     this.display = value;
@@ -119,7 +137,7 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
 
   /**
    * Convert object to human readable output
-   * @param value Display output
+   * @param value Display input
    */
   private humanizeObject(value: object) {
     this.setLoading();
@@ -134,7 +152,7 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
 
   /**
    * Convert blob to human readable output
-   * @param value Display output
+   * @param value Display input
    */
   private humanizeBlob(value: Blob) {
     this.setLoading();
@@ -151,10 +169,14 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
     reader.readAsText(value);
   }
 
+  /**
+   * Convert observable to human readable output
+   * @param value Display input
+   */
   private humanizeObservable(
     value: Observable<AbstractModel | AbstractModel[]>
   ) {
-    this.display = this.loadingText;
+    this.setLoading();
     value.pipe(takeUntil(this.unsubscribe)).subscribe(
       (models) => {
         if (!models) {
@@ -173,7 +195,7 @@ export class RenderFieldComponent extends WithUnsubscribe() implements OnInit {
 
   /**
    * Convert array to human readable output
-   * @param value Display output
+   * @param value Display input
    */
   private humanizeArray(value: ModelView[]) {
     if (value.length > 0) {
