+++ src/app/models/Dataset.ts
@@ -5,7 +5,6 @@ import {
   Id,
   Param,
 } from "@interfaces/apiInterfaces";
-import { Observable } from "rxjs";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
